+++ src/main/kotlin/dev/voqal/ide/VoqalIcons.kt
@@ -36,6 +36,7 @@ object VoqalIcons {
         val anthropic: Icon = IconLoader.getIcon("/icons/compute/anthropic.svg", VoqalIcons::class.java)
         val deepseek: Icon = IconLoader.getIcon("/icons/compute/deepseek.svg", VoqalIcons::class.java)
         val fireworks: Icon = IconLoader.getIcon("/icons/compute/fireworks.svg", VoqalIcons::class.java)
+        val sambanova: Icon = IconLoader.getIcon("/icons/compute/sambanova.svg", VoqalIcons::class.java)
     }
 
     val DARK_RED = JBColor(JBColor.PINK, Color(0x5E3838))